First approach typescript sequelize
