--- app/models/users.js
@@ -1,14 +1,0 @@
-module.exports = (sequelize, DataTypes) => {
-  const User = sequelize.define(
-    'users',
-    {
-      username: {
-        type: DataTypes.STRING
-      }
-    },
-    {
-      timestamps: false
-    }
-  );
-  return User;
-};
