+++ package.json
@@ -67,7 +67,7 @@
     "pg": "^7.4.1",
     "request-promise": "^4.2.4",
     "rollbar": "^2.3.9",
-    "sequelize": "^4.34.0",
+    "sequelize": "^5.15.0",
     "umzug": "^2.1.0",
     "winston": "^2.4.0",
     "winston-daily-rotate-file": "^1.7.2"
@@ -85,6 +85,7 @@
     "@types/pg": "^7.4.14",
     "@types/request-promise": "^4.1.44",
     "@types/sequelize": "^4.28.3",
+    "@types/umzug": "^2.2.2",
     "babel": "6.23.0",
     "babel-core": "6.26.0",
     "babel-eslint": "^8.2.2",
