+++ server.ts
@@ -1,9 +1,7 @@
-
-
 import app from './app';
 import Rollbar from 'rollbar';
-import migrationsManager from './migrations';
-import config from './config';
+import * as migrationsManager from './migrations';
+import * as config from './config';
 import logger from './app/logger';
 
 const port = config.common.api.port || 8080;
